+++ Backend/app/rider-platform/rider-app/Main/src/Domain/Action/UI/MultimodalConfirm.hs
@@ -73,14 +73,12 @@ import qualified Lib.JourneyModule.Types as JMTypes
 import qualified Lib.JourneyModule.Utils as JLU
 import qualified Lib.Payment.Domain.Types.PaymentOrder as DOrder
 import qualified Lib.Payment.Storage.Queries.PaymentOrder as QOrder
-import qualified SharedLogic.CreateFareForMultiModal as SMMFRFS
 import qualified Storage.CachedQueries.IntegratedBPPConfig as QIBC
 import qualified Storage.CachedQueries.Merchant.MerchantOperatingCity as CQMOC
 import qualified Storage.CachedQueries.OTPRest.OTPRest as OTPRest
 import qualified Storage.Queries.BecknConfig as QBC
 import qualified Storage.Queries.Estimate as QEstimate
 import qualified Storage.Queries.FRFSQuote as QFRFSQuote
-import qualified Storage.Queries.FRFSQuote as QQuote
 import Storage.Queries.FRFSSearch as QFRFSSearch
 import qualified Storage.Queries.FRFSTicketBokingPayment as QFRFSTicketBookingPayment
 import Storage.Queries.FRFSTicketBooking as QFRFSTicketBooking
@@ -213,69 +211,19 @@ postMultimodalPaymentUpdateOrder ::
   )
 postMultimodalPaymentUpdateOrder (mbPersonId, _merchantId) journeyId req = do
   personId <- fromMaybeM (InvalidRequest "Invalid person id") mbPersonId
-  person <- QP.findById personId >>= fromMaybeM (InvalidRequest "Person not found")
-  frfsBookingsArr <- QFRFSTicketBooking.findAllByJourneyIdCond (Just journeyId)
-  frfsBookingWithUpdatedPriceAndQty <-
-    mapM
-      ( \ticketBooking -> do
-          quote <- QQuote.findById ticketBooking.quoteId >>= fromMaybeM (FRFSQuoteNotFound $ show ticketBooking.quoteId)
-          let quantityRational = fromIntegral req.quantity :: Rational
-          let oldQuantityRational = fromIntegral ticketBooking.quantity :: Rational
-          let childTicketQuantityRational = fromIntegral req.childTicketQuantity :: Rational
-          let childPrice = maybe (getHighPrecMoney quote.price.amount) (\p -> getHighPrecMoney p.amount) quote.childPrice
-          let amountToBeUpdated = ((safeDiv (getHighPrecMoney ticketBooking.price.amount) oldQuantityRational) * quantityRational) + (childPrice * childTicketQuantityRational)
-          let totalPrice =
-                Price
-                  { amount = HighPrecMoney $ amountToBeUpdated,
-                    amountInt = Money $ roundToIntegral amountToBeUpdated,
-                    currency = ticketBooking.price.currency
-                  }
-          let fRFSTicketBooking =
-                ticketBooking
-                  { DFRFSB.quantity = req.quantity,
-                    DFRFSB.childTicketQuantity = Just req.childTicketQuantity,
-                    DFRFSB.finalPrice = Just totalPrice,
-                    DFRFSB.estimatedPrice = totalPrice,
-                    DFRFSB.price = totalPrice
-                  }
-          QFRFSTicketBooking.updateByPrimaryKey fRFSTicketBooking
-          pure $ ticketBooking{price = totalPrice, quantity = req.quantity}
-      )
-      frfsBookingsArr
-  frfsBookingsPaymentArr <- mapM (QFRFSTicketBookingPayment.findAllTicketBookingId . (.id)) frfsBookingWithUpdatedPriceAndQty
-  let paymentType = TPayment.FRFSMultiModalBooking
-  (_vendorSplitDetails, amountUpdated) <- SMMFRFS.createVendorSplitFromBookings frfsBookingWithUpdatedPriceAndQty _merchantId person.merchantOperatingCityId paymentType
-  isSplitEnabled <- TPayment.getIsSplitEnabled _merchantId person.merchantOperatingCityId Nothing paymentType
-  let splitDetails = TPayment.mkUnaggregatedSplitSettlementDetails isSplitEnabled amountUpdated _vendorSplitDetails
-  let flattenedPayments = concat frfsBookingsPaymentArr
-  let mbPaymentOrderId = paymentOrderId <$> listToMaybe flattenedPayments
-  case mbPaymentOrderId of
+  mbUpdatedOrder <- JLU.postMultimodalPaymentUpdateOrderUtil personId _merchantId (Just journeyId) req.quantity req.childTicketQuantity
+  case mbUpdatedOrder of
     Nothing ->
       return $
         ApiTypes.UpdatePaymentOrderResp
           { sdkPayload = Nothing
           }
-    Just paymentOrderId -> do
-      order <- QOrder.findById paymentOrderId >>= fromMaybeM (PaymentOrderNotFound paymentOrderId.getId)
-      let updateReq =
-            KT.OrderUpdateReq
-              { amount = amountUpdated,
-                orderShortId = order.shortId.getShortId,
-                splitSettlementDetails = splitDetails
-              }
-      _ <- TPayment.updateOrder person.merchantId person.merchantOperatingCityId Nothing TPayment.FRFSMultiModalBooking (Just person.id.getId) person.clientSdkVersion updateReq
-      QOrder.updateAmount order.id amountUpdated
-      let updatedOrder :: DOrder.PaymentOrder
-          updatedOrder = order {DOrder.amount = amountUpdated}
-      sdkPayload <- buildUpdateOrderSDKPayload amountUpdated updatedOrder
+    Just updatedOrder -> do
+      sdkPayload <- buildUpdateOrderSDKPayload updatedOrder.amount updatedOrder
       return $
         ApiTypes.UpdatePaymentOrderResp
           { sdkPayload = sdkPayload
           }
-  where
-    safeDiv :: (Eq a, Fractional a) => a -> a -> a
-    safeDiv x 0 = x
-    safeDiv x y = x / y
 
 buildUpdateOrderSDKPayload :: EncFlow m r => HighPrecMoney -> DOrder.PaymentOrder -> m (Maybe Juspay.SDKPayloadDetails)
 buildUpdateOrderSDKPayload amount order = do
